Use telega-username as the face for tracking mode-line notifications
When adding a buffer for notifications with the tracking package, one can specify a list of faces to use for it in the modeline.  This is useful not only for appearance, but also because the face can be used to prioritize the notification, altering the order in which it is shown (via tracking-faces-priorities).  I think telega-username is a good candidate for this, at least as a default.